requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
openai>=1.0.0
anthropic>=0.7.0
ollama>=0.1.0
openai-whisper>=20230314
python-dotenv>=1.0.0
Pillow>=10.0.0